Move low-level Metric/Family objects to metrics.py (keep re-exports in core)
